Move rgp into extensions directory
