+++ epp-client/src/tests/se.rs
@@ -21,8 +21,6 @@ mod request {
     use crate::domain::delete::DomainDelete;
     use crate::domain::info::DomainInfo;
     use crate::domain::renew::DomainRenew;
-    use crate::domain::rgp::report::RgpRestoreReport;
-    use crate::domain::rgp::request::RgpRestoreRequest;
     use crate::domain::transfer::DomainTransferApprove;
     use crate::domain::transfer::DomainTransferCancel;
     use crate::domain::transfer::DomainTransferQuery;
@@ -31,6 +29,8 @@ mod request {
     use crate::domain::update::DomainAddRemove;
     use crate::domain::update::DomainChangeInfo;
     use crate::domain::update::DomainUpdate;
+    use crate::extensions::rgp::report::RgpRestoreReport;
+    use crate::extensions::rgp::request::RgpRestoreRequest;
     use crate::hello::Hello;
     use crate::host::check::HostCheck;
     use crate::host::create::HostCreate;
